requirements.txt
cryptography
argparse
pycrypto